requirements.txt
fastapi
google-api-python-client
Jinja2
networkx
protobuf
pydantic
pyvis
requests
SQLAlchemy
uvicorn
google-generativeai
google-cloud-asset
celery
redis
google-auth
croniter
